post-deploy.bash: make ln overwrite a symlink to a directory
The post_deploy.bash effectively did:

  ln -sfv [DIRECTORY-TARGET] [LINK-NAME]

This works fine if [LINK-NAME] did not exist, but on subsequent
runs, [LINK-NAME] will be dereferenced, and that parameter
will be treated as a directory.  This causes a circular
symlink to be created in that directory.  Adding the -n
parameter to ln tells it not to dereference [LINK-NAME] when
[LINK-NAME] is a symlink to a directory.