Brighton stylesheet asset issue
This is an interestingly subtle problem, affecting images referenced
from stylesheets. Because the server has no natural way of knowing that a
stylesheet request is via the admin site, it’s hard to change the image
paths to be also via the admin site, hence we still get mixed content
warnings even after the previous round of changes.

We pass a query string parameter to the stylesheet requests, which
ought to make it possible to fix this.